% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quality.R
\name{analyze_vo2_workload_relationship}
\alias{analyze_vo2_workload_relationship}
\title{Analyze VO2/Workload Relationship}
\usage{
analyze_vo2_workload_relationship(breaths, expected_slope, modality)
}
\arguments{
\item{breaths}{Breath data tibble}

\item{expected_slope}{Expected slope (mL/min/W)}

\item{modality}{Exercise modality}
}
\value{
List with actual_slope, deviation_pct, acceptable, r_squared, stage_details
}
\description{
Analyze VO2/Workload Relationship
}
\keyword{internal}
