% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quality.R
\name{assess_baseline_stability}
\alias{assess_baseline_stability}
\title{Assess Baseline Stability}
\usage{
assess_baseline_stability(breaths)
}
\arguments{
\item{breaths}{Breath data tibble}
}
\value{
List with cv (coefficient of variation of resting VO2)
}
\description{
Assess Baseline Stability
}
\keyword{internal}
