% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-average.R
\name{average}
\alias{average}
\title{Average CPET Data}
\usage{
average(x, ...)
}
\arguments{
\item{x}{A CpetData object}

\item{...}{Additional arguments passed to methods, including:
\describe{
\item{method}{Averaging method: "time" (time-based windows), "breath"
(fixed number of breaths), or "rolling" (rolling/moving average)}
\item{window}{Window size: seconds for "time" method, number of breaths
for "breath" method, seconds for "rolling" method}
}}
}
\value{
A new CpetData object with averaged data and is_averaged = TRUE
}
\description{
Apply averaging to breath-by-breath CPET data.
}
\examples{
file <- system.file("extdata", "example_cosmed.xlsx", package = "cardiometR")
data <- read_cpet(file, quiet = TRUE)
averaged <- average(data, method = "time", window = 30)
averaged@is_averaged

}
