% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-average.R
\name{average_by_breath}
\alias{average_by_breath}
\title{Average Breaths by Fixed Number of Breaths}
\usage{
average_by_breath(breaths, window)
}
\arguments{
\item{breaths}{Breath-by-breath data tibble}

\item{window}{Number of breaths per average}
}
\value{
Averaged tibble
}
\description{
Groups consecutive breaths and calculates mean values.
}
\keyword{internal}
