% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-average.R
\name{average_rolling}
\alias{average_rolling}
\title{Rolling/Moving Average}
\usage{
average_rolling(breaths, window)
}
\arguments{
\item{breaths}{Breath-by-breath data tibble}

\item{window}{Window size in seconds}
}
\value{
Tibble with rolling averages
}
\description{
Applies a rolling mean with specified window size.
}
\keyword{internal}
