% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quality.R
\name{calculate_exercise_criteria_score}
\alias{calculate_exercise_criteria_score}
\title{Calculate Exercise Criteria Score}
\usage{
calculate_exercise_criteria_score(criteria)
}
\arguments{
\item{criteria}{ExerciseQualityCriteria object}
}
\value{
Numeric score 0-100
}
\description{
Calculate Exercise Criteria Score
}
\keyword{internal}
