% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quality.R
\name{calculate_exercise_duration}
\alias{calculate_exercise_duration}
\title{Calculate Exercise Duration}
\usage{
calculate_exercise_duration(breaths)
}
\arguments{
\item{breaths}{Breath data tibble}
}
\value{
List with duration_s and warmup_s
}
\description{
Calculate Exercise Duration
}
\keyword{internal}
