% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quality.R
\name{calculate_signal_quality_score}
\alias{calculate_signal_quality_score}
\title{Calculate Signal Quality Score}
\usage{
calculate_signal_quality_score(pct_aberrant, pct_missing_hr, baseline_cv)
}
\arguments{
\item{pct_aberrant}{Percentage of aberrant breaths}

\item{pct_missing_hr}{Percentage of missing HR (or NULL)}

\item{baseline_cv}{Baseline VO2 CV (or NULL)}
}
\value{
Numeric score 0-100
}
\description{
Calculate Signal Quality Score
}
\keyword{internal}
