% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-peaks.R
\name{check_vo2_plateau}
\alias{check_vo2_plateau}
\title{Calculate VO2 Plateau}
\usage{
check_vo2_plateau(x, threshold = 150)
}
\arguments{
\item{x}{A CpetData object with stages}

\item{threshold}{Maximum VO2 increase to consider a plateau (default 150 mL/min)}
}
\value{
List with plateau status and details
}
\description{
Checks if a VO2 plateau was achieved (< 150 mL/min increase
in final stage despite increased workload).
}
\examples{
file <- system.file("extdata", "example_cosmed.xlsx", package = "cardiometR")
data <- extract_stages(read_cpet(file, quiet = TRUE))
check_vo2_plateau(data)

}
