% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-graphs.R
\name{cleanup_temp_files}
\alias{cleanup_temp_files}
\title{Clean Up Temporary Files}
\usage{
cleanup_temp_files(files)
}
\arguments{
\item{files}{List of file paths to remove}
}
\description{
Clean Up Temporary Files
}
\keyword{internal}
