% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-report.R
\name{cpet_import_report}
\alias{cpet_import_report}
\title{Get the import report attached to a CpetData object}
\usage{
cpet_import_report(x)
}
\arguments{
\item{x}{A \link{CpetData} object.}
}
\value{
A \link{CpetImportReport}, or \code{NULL} when the object was not produced by
\code{\link[=read_cpet]{read_cpet()}}.
}
\description{
Get the import report attached to a CpetData object
}
\examples{
file <- system.file("extdata", "example_cosmed.xlsx", package = "cardiometR")
data <- read_cpet(file, quiet = TRUE)
report <- cpet_import_report(data)
report@columns[report@columns$status == "mapped", c("source", "canonical")]

}
\seealso{
\code{\link[=read_cpet]{read_cpet()}}, \code{\link[=preview_cpet_columns]{preview_cpet_columns()}}
}
