% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quality.R
\name{criterion_item}
\alias{criterion_item}
\title{Create Criterion Item}
\usage{
criterion_item(met, label, value, lang)
}
\arguments{
\item{met}{Logical or NULL indicating if criterion was met}

\item{label}{Criterion label}

\item{value}{Achieved value}

\item{lang}{Language code}
}
\value{
Shiny tag
}
\description{
Create Criterion Item
}
\keyword{internal}
