% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stages.R
\name{detect_power_increment}
\alias{detect_power_increment}
\title{Detect Power Increment Used in Protocol}
\usage{
detect_power_increment(power)
}
\arguments{
\item{power}{Power values}
}
\value{
Detected power increment
}
\description{
Detect Power Increment Used in Protocol
}
\keyword{internal}
