% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stages.R
\name{extract_stages_by_time}
\alias{extract_stages_by_time}
\title{Extract Stages by Time Windows}
\usage{
extract_stages_by_time(breaths, stage_duration)
}
\arguments{
\item{breaths}{Breath data tibble}

\item{stage_duration}{Stage duration in seconds}
}
\value{
Tibble with stage annotations
}
\description{
Extract Stages by Time Windows
}
\keyword{internal}
