% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-dialects.R
\name{find_dialect_fragment}
\alias{find_dialect_fragment}
\title{Find a shared fragment such as \verb{_core.yml}}
\usage{
find_dialect_fragment(name)
}
\arguments{
\item{name}{Fragment name without extension.}
}
\value{
A path, or \code{NA} when the fragment does not exist.
}
\description{
Find a shared fragment such as \verb{_core.yml}
}
\keyword{internal}
