% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find Peak Values}
\usage{
find_peaks(x, ...)
}
\arguments{
\item{x}{A CpetData object}

\item{...}{Additional arguments passed to methods, including:
\describe{
\item{averaging}{Seconds to average for peak determination (default 30)}
}}
}
\value{
A PeakValues S7 object containing peak VO2, HR, VE, RER, etc.
}
\description{
Determine peak/maximal values from CPET data.
}
\details{
Peak values are determined using a rolling average of the specified window
size to reduce breath-by-breath variability. The recommended default is
30 seconds per ATS/ACCP guidelines.
}
\examples{
file <- system.file("extdata", "example_cosmed.xlsx", package = "cardiometR")
data <- read_cpet(file, quiet = TRUE)
peaks <- find_peaks(data, averaging = 30)
peaks@vo2_peak
peaks@vo2_kg_peak

}
