% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-typst.R
\name{format_modality}
\alias{format_modality}
\title{Format Modality}
\usage{
format_modality(modality, language = "en")
}
\arguments{
\item{modality}{Character: "cycling" or "treadmill"}

\item{language}{Language code ("en" or "fr")}
}
\value{
Formatted string
}
\description{
Format Modality
}
\keyword{internal}
