% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quality.R
\name{generate_data_quality_recommendations}
\alias{generate_data_quality_recommendations}
\title{Generate Data Quality Recommendations}
\usage{
generate_data_quality_recommendations(
  pct_aberrant,
  pct_missing_hr,
  pct_missing_power,
  baseline_cv,
  drift_detected
)
}
\arguments{
\item{pct_aberrant}{Percentage of aberrant breaths}

\item{pct_missing_hr}{Percentage of missing HR}

\item{pct_missing_power}{Percentage of missing power}

\item{baseline_cv}{Baseline VO2 CV}

\item{drift_detected}{Whether calibration drift was detected}
}
\value{
List of recommendation strings
}
\description{
Generate Data Quality Recommendations
}
\keyword{internal}
