% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quality.R
\name{generate_quality_summary}
\alias{generate_quality_summary}
\title{Generate Quality Summary Text}
\usage{
generate_quality_summary(
  exercise_criteria,
  protocol_quality,
  data_quality,
  grade
)
}
\arguments{
\item{exercise_criteria}{ExerciseQualityCriteria object}

\item{protocol_quality}{ProtocolQuality object}

\item{data_quality}{DataQualityReport object}

\item{grade}{Overall grade}
}
\value{
Character summary
}
\description{
Generate Quality Summary Text
}
\keyword{internal}
