% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normative-data.R
\name{get_age_group}
\alias{get_age_group}
\title{Get Age Group Category}
\usage{
get_age_group(age)
}
\arguments{
\item{age}{Age in years}
}
\value{
Character age group label
}
\description{
Get Age Group Category
}
\keyword{internal}
