% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normative-data.R
\name{get_age_midpoint}
\alias{get_age_midpoint}
\title{Get Age Group Midpoint}
\usage{
get_age_midpoint(age_group)
}
\arguments{
\item{age_group}{Age group category}
}
\value{
Numeric midpoint of age range
}
\description{
Get Age Group Midpoint
}
\keyword{internal}
