% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normative-data.R
\name{list_normative_categories}
\alias{list_normative_categories}
\title{Get All Available Normative Categories}
\usage{
list_normative_categories()
}
\value{
A tibble with sport, level, sex, and typical VO2max values
}
\description{
Returns a data frame of all available normative data categories
with typical VO2max values for quick reference.
}
\examples{
list_normative_categories()

}
