% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quality.R
\name{metric_row}
\alias{metric_row}
\title{Create Metric Row}
\usage{
metric_row(label, value, status = "secondary", bold = FALSE)
}
\arguments{
\item{label}{Metric label}

\item{value}{Metric value}

\item{status}{Bootstrap color (success, warning, danger, etc.)}

\item{bold}{Whether to bold the value}
}
\value{
Shiny tag
}
\description{
Create Metric Row
}
\keyword{internal}
