% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_participant.R
\name{mod_participant_server}
\alias{mod_participant_server}
\title{Participant Module Server}
\usage{
mod_participant_server(id, language, cpet_data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Reactive language value.}

\item{cpet_data}{Reactive CpetData object from upload module.}
}
\value{
A list with reactive values:
\itemize{
\item \code{participant}: Reactive Participant object with user edits.
\item \code{prediction_source}: Reactive prediction equation source ("jones" or "prefaut").
}
}
\description{
Participant Module Server
}
\keyword{internal}
