% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_participant.R
\name{mod_participant_ui}
\alias{mod_participant_ui}
\title{Participant Module UI}
\usage{
mod_participant_ui(id, language = "en")
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Language code ("en" or "fr").}
}
\value{
A Shiny UI element.
}
\description{
Participant Module UI
}
\keyword{internal}
