% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{mod_plots_ui}
\alias{mod_plots_ui}
\title{Plots Module UI}
\usage{
mod_plots_ui(id, language = "en", secondary_id = NULL)
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Language code ("en" or "fr").}
}
\value{
A Shiny UI element.
}
\description{
Plots Module UI
}
\keyword{internal}
