% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quality.R
\name{mod_quality_server}
\alias{mod_quality_server}
\title{Quality Assessment Module Server}
\usage{
mod_quality_server(id, language, cpet_data, analysis)
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Reactive language value.}

\item{cpet_data}{Reactive CpetData object from upload module.}

\item{analysis}{Reactive CpetAnalysis object from results module.}
}
\value{
A list with reactive values:
\itemize{
\item \code{quality}: Reactive QualityAssessment object.
}
}
\description{
Quality Assessment Module Server
}
\keyword{internal}
