% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{Report Module Server}
\usage{
mod_report_server(id, language, analysis)
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Reactive language value.}

\item{analysis}{Reactive CpetAnalysis object from results module.}
}
\description{
Report Module Server
}
\keyword{internal}
