% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_ui}
\alias{mod_report_ui}
\title{Report Module UI}
\usage{
mod_report_ui(id, language = "en")
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Language code ("en" or "fr").}
}
\value{
A Shiny UI element.
}
\description{
Report Module UI
}
\keyword{internal}
