% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_results_ui}
\alias{mod_results_ui}
\title{Results Module UI}
\usage{
mod_results_ui(id, language = "en")
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Language code ("en" or "fr").}
}
\value{
A Shiny UI element.
}
\description{
Results Module UI
}
\keyword{internal}
