% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_thresholds.R
\name{mod_thresholds_ui}
\alias{mod_thresholds_ui}
\title{Threshold Review Module UI}
\usage{
mod_thresholds_ui(id, language = "en")
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Language code ("en" or "fr").}
}
\value{
A Shiny UI element.
}
\description{
Threshold Review Module UI
}
\keyword{internal}
