% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_server}
\alias{mod_upload_server}
\title{Upload Module Server}
\usage{
mod_upload_server(id, language)
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Reactive language value.}
}
\value{
A list with reactive values:
\itemize{
\item \code{cpet_data}: Reactive CpetData object (or NULL).
\item \code{validation}: Reactive ValidationReport object (or NULL).
}
}
\description{
Upload Module Server
}
\keyword{internal}
