% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_ui}
\alias{mod_upload_ui}
\title{Upload Module UI}
\usage{
mod_upload_ui(id, language = "en")
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Language code ("en" or "fr").}
}
\value{
A Shiny UI element.
}
\description{
Upload Module UI
}
\keyword{internal}
