% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normative-data.R
\name{normative_lookup}
\alias{normative_lookup}
\title{Normative Data Lookup}
\usage{
normative_lookup(sport, level, sex, age_group)
}
\arguments{
\item{sport}{Sport type}

\item{level}{Competitive level}

\item{sex}{Sex}

\item{age_group}{Age group category}
}
\value{
List with normative values and citations
}
\description{
Normative Data Lookup
}
\keyword{internal}
