% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_cosmed_breaths}
\alias{parse_cosmed_breaths}
\title{Parse Breath-by-Breath Data from COSMED Export}
\usage{
parse_cosmed_breaths(raw)
}
\arguments{
\item{raw}{Raw data frame from readxl}
}
\value{
A tibble of breath-by-breath data
}
\description{
Parse Breath-by-Breath Data from COSMED Export
}
\keyword{internal}
