% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_cosmed_metadata}
\alias{parse_cosmed_metadata}
\title{Parse Metadata from COSMED Export}
\usage{
parse_cosmed_metadata(raw)
}
\arguments{
\item{raw}{Raw data frame from readxl}
}
\value{
A CpetMetadata S7 object
}
\description{
Parse Metadata from COSMED Export
}
\keyword{internal}
