% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_cosmed_participant}
\alias{parse_cosmed_participant}
\title{Parse Participant Information from COSMED Export}
\usage{
parse_cosmed_participant(raw)
}
\arguments{
\item{raw}{Raw data frame from readxl}
}
\value{
A Participant S7 object
}
\description{
Parse Participant Information from COSMED Export
}
\keyword{internal}
