% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{PeakValues}
\alias{PeakValues}
\title{Peak Values Class}
\usage{
PeakValues(
  vo2_peak = integer(0),
  vo2_kg_peak = integer(0),
  ve_peak = integer(0),
  rer_peak = integer(0),
  averaging_s = integer(0),
  hr_peak = integer(0),
  power_peak = integer(0)
)
}
\arguments{
\item{vo2_peak}{Peak oxygen uptake in mL/min}

\item{vo2_kg_peak}{Peak oxygen uptake relative to body weight in mL/kg/min}

\item{ve_peak}{Peak minute ventilation in L/min}

\item{rer_peak}{Peak respiratory exchange ratio}

\item{averaging_s}{Averaging window used to determine peaks (seconds)}

\item{hr_peak}{Optional peak heart rate in bpm}

\item{power_peak}{Optional peak power output in watts}
}
\value{
A PeakValues S7 object
}
\description{
Container for peak/maximal values from a CPET test.
}
