% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normative-data.R
\name{percentile_from_z}
\alias{percentile_from_z}
\title{Percentile from Z-Score}
\usage{
percentile_from_z(z)
}
\arguments{
\item{z}{Z-score (scalar or vector).}
}
\value{
Percentile(s) on a 0-100 scale.
}
\description{
Percentile from Z-Score
}
