% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heart_rate}
\alias{plot_heart_rate}
\title{Plot Heart Rate Response}
\usage{
plot_heart_rate(x, x_axis = "time", show_zones = FALSE, language = "en")
}
\arguments{
\item{x}{A CpetData object}

\item{x_axis}{X-axis variable: "time" or "vo2"}

\item{show_zones}{Show training zones based on peak HR}

\item{language}{Language for labels}
}
\value{
A ggplot2 object
}
\description{
Plots heart rate against time or VO2 with optional target zones.
}
\examples{
\dontrun{
file <- system.file("extdata", "example_cosmed.xlsx", package = "cardiometR")
plot_heart_rate(read_cpet(file), show_zones = TRUE)
}

}
