% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{print.CpetImportReport}
\alias{print.CpetImportReport}
\title{Print a summary of what an import did}
\usage{
## S7 method for class <cardiometR::CpetImportReport>
print(x, ...)
}
\arguments{
\item{x}{A \link{CpetImportReport}.}

\item{...}{Unused.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print a summary of what an import did
}
\keyword{internal}
