% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods-print.R
\name{QualityAssessment}
\alias{QualityAssessment}
\alias{print.QualityAssessment}
\title{Quality Assessment Class}
\usage{
QualityAssessment(
  exercise_criteria = ExerciseQualityCriteria(),
  protocol_quality = ProtocolQuality(),
  data_quality = DataQualityReport(),
  overall_grade = character(0),
  overall_score = integer(0),
  test_interpretable = logical(0),
  summary_text = character(0)
)
}
\arguments{
\item{exercise_criteria}{ExerciseQualityCriteria S7 object}

\item{protocol_quality}{ProtocolQuality S7 object}

\item{data_quality}{DataQualityReport S7 object}

\item{overall_grade}{Overall test grade: "A", "B", "C", "D", "F"}

\item{overall_score}{Combined quality score 0-100}

\item{test_interpretable}{Logical indicating if test results can be reliably interpreted}

\item{summary_text}{Brief text summary of quality assessment}
}
\value{
A QualityAssessment S7 object
}
\description{
Complete quality assessment combining exercise criteria, protocol quality,
and data quality into a unified report.
}
