% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-dialects.R
\name{read_dialect_file}
\alias{read_dialect_file}
\title{Read one dialect YAML file, with caching}
\usage{
read_dialect_file(path)
}
\arguments{
\item{path}{Path to a \code{.yml} file.}
}
\value{
The parsed list.
}
\description{
Read one dialect YAML file, with caching
}
\keyword{internal}
