% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-cache.R
\name{session_cache_default_path}
\alias{session_cache_default_path}
\title{Default path for the latest session snapshot}
\usage{
session_cache_default_path()
}
\description{
Default path for the latest session snapshot
}
\keyword{internal}
