% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stages.R
\name{summarize_last_n_seconds}
\alias{summarize_last_n_seconds}
\title{Summarize Last N Seconds of Stage}
\usage{
summarize_last_n_seconds(df, seconds)
}
\arguments{
\item{df}{Stage data}

\item{seconds}{Number of seconds from end of stage}
}
\value{
Summary tibble
}
\description{
Summarize Last N Seconds of Stage
}
\keyword{internal}
