% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stages.R
\name{summarize_mean}
\alias{summarize_mean}
\title{Summarize Stage by Mean}
\usage{
summarize_mean(df)
}
\arguments{
\item{df}{Stage data}
}
\value{
Summary tibble
}
\description{
Summarize Stage by Mean
}
\keyword{internal}
