% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stages.R
\name{summarize_peak}
\alias{summarize_peak}
\title{Summarize Stage by Peak Values}
\usage{
summarize_peak(df)
}
\arguments{
\item{df}{Stage data}
}
\value{
Summary tibble
}
\description{
Summarize Stage by Peak Values
}
\keyword{internal}
