% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stages.R
\name{summarize_stage_numeric}
\alias{summarize_stage_numeric}
\title{Summarize Stage by Mean}
\usage{
summarize_stage_numeric(df, stat = c("mean", "max"))
}
\arguments{
\item{df}{Stage data}
}
\value{
Summary tibble
}
\description{
Summarize Stage by Mean
}
\keyword{internal}
