% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{validate_language}
\alias{validate_language}
\title{Validate Language Code}
\usage{
validate_language(language)
}
\arguments{
\item{language}{Language code to validate}
}
\value{
Valid language code ("en" or "fr")
}
\description{
Validate Language Code
}
\keyword{internal}
