% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-economy.R
\name{calculate_gross_efficiency}
\alias{calculate_gross_efficiency}
\title{Calculate Gross Efficiency}
\usage{
calculate_gross_efficiency(vo2_ml, power_w, rer = 1)
}
\arguments{
\item{vo2_ml}{VO2 in mL/min}

\item{power_w}{Power output in watts}

\item{rer}{Respiratory exchange ratio (default 1.0 for high intensity)}
}
\value{
Gross efficiency as percentage
}
\description{
Calculates cycling gross efficiency (GE\%) from VO2 and power output.
Gross efficiency represents the ratio of mechanical work output to
metabolic energy expenditure.
}
\details{
Formula: GE = (Power output / Energy expenditure) * 100

The energy equivalent of oxygen varies with RER:
\itemize{
\item RER 0.70: 4.69 kcal/L O2 (pure fat oxidation)
\item RER 1.00: 5.05 kcal/L O2 (pure carbohydrate oxidation)
}

Typical values:
\itemize{
\item Elite cyclists: 22-25\%
\item Trained cyclists: 19-22\%
\item Recreational: 17-20\%
}
}
\examples{
# At 3000 mL/min VO2, 250W, RER 1.0
calculate_gross_efficiency(3000, 250, 1.0)  # ~20.9\%

}
\references{
Coyle EF, Sidossis LS, Horowitz JF, Beltz JD. Cycling efficiency is related
to the percentage of type I muscle fibers. Med Sci Sports Exerc. 1992;24(7):782-788.
}
