% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{calculate_predicted_values}
\alias{calculate_predicted_values}
\title{Calculate Predicted Values}
\usage{
calculate_predicted_values(participant)
}
\arguments{
\item{participant}{Participant object}
}
\value{
Named list of predicted values
}
\description{
Calculates age/sex-specific predicted maximal values.
Uses standard prediction equations.
}
\keyword{internal}
