% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiometR-package.R
\docType{package}
\name{cardiometR-package}
\alias{cardiometR}
\alias{cardiometR-package}
\title{cardiometR: Cardiopulmonary Exercise Testing Analysis with S7 Classes}
\description{
Import, analyze, and report cardiopulmonary exercise testing (CPET) data from COSMED Quark CPET metabolic carts. Provides S7 classes for type-safe physiological data structures, peak and ventilatory threshold detection following ATS/ACCP guidelines with operator review, test-quality assessment against ACSM criteria, normative comparison, an interactive Shiny application, and bilingual (English/French) clinical PDF reports rendered with Typst.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jotremblay/cardiometR}
  \item \url{https://jotremblay.github.io/cardiometR/}
  \item Report bugs at \url{https://github.com/jotremblay/cardiometR/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Tremblay \email{jonathan.tremblay@umontreal.ca}

Authors:
\itemize{
  \item Jonathan Tremblay \email{jonathan.tremblay@umontreal.ca}
}

}
\keyword{internal}
