% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiometR-package.R
\docType{package}
\name{cardiometR-package}
\alias{cardiometR}
\alias{cardiometR-package}
\title{cardiometR: Cardiopulmonary Exercise Testing Analysis with S7 Classes}
\description{
Import, process, and report cardiopulmonary exercise testing (CPET) data from COSMED Quark CPET metabolic carts. Uses S7 classes for type-safe data structures. Includes a Shiny application for interactive analysis and bilingual (English/French) PDF report generation via Typst.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jotremblay/cardiometR}
  \item Report bugs at \url{https://github.com/jotremblay/cardiometR/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Tremblay \email{jonathan.tremblay@umontreal.ca}

}
\keyword{internal}
