% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validate.R
\name{check_data_completeness}
\alias{check_data_completeness}
\title{Check Data Completeness}
\usage{
check_data_completeness(breaths)
}
\arguments{
\item{breaths}{Breath data tibble}
}
\value{
List with warnings, info
}
\description{
Check Data Completeness
}
\keyword{internal}
