% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validate.R
\name{check_maximal_effort}
\alias{check_maximal_effort}
\title{Check Maximal Effort Indicators}
\usage{
check_maximal_effort(breaths, age)
}
\arguments{
\item{breaths}{Breath data tibble}

\item{age}{Participant age}
}
\value{
List with info messages about effort level
}
\description{
Check Maximal Effort Indicators
}
\keyword{internal}
