% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validate.R
\name{check_physiological_ranges}
\alias{check_physiological_ranges}
\title{Check Physiological Plausibility}
\usage{
check_physiological_ranges(breaths, participant)
}
\arguments{
\item{breaths}{Breath data tibble}

\item{participant}{Participant S7 object}
}
\value{
List with errors, warnings, info
}
\description{
Check Physiological Plausibility
}
\keyword{internal}
