% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-dialects.R
\name{clear_dialect_cache}
\alias{clear_dialect_cache}
\title{Clear the dialect cache}
\usage{
clear_dialect_cache()
}
\value{
Invisibly \code{NULL}.
}
\description{
Useful when editing a dialect file during a session.
}
\keyword{internal}
