% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{CpetMetadata}
\alias{CpetMetadata}
\title{CPET Metadata Class}
\usage{
CpetMetadata(
  test_date = (function (.data = double()) 
 {
     .Date(.data)
 })(),
  device = character(0),
  protocol = character(0),
  calibration_date = (function (.data = double()) 
 {
     .Date(.data)
 })(),
  temperature_c = integer(0),
  pressure_mmhg = integer(0),
  humidity_pct = integer(0),
  technician = character(0)
)
}
\arguments{
\item{test_date}{Date of the CPET test}

\item{device}{Device/metabolic cart name (e.g., "COSMED Quark CPET")}

\item{protocol}{Exercise protocol name (e.g., "Bruce", "Ramp")}

\item{calibration_date}{Optional date of last calibration}

\item{temperature_c}{Optional ambient temperature in Celsius}

\item{pressure_mmhg}{Optional barometric pressure in mmHg}

\item{humidity_pct}{Optional relative humidity percentage}

\item{technician}{Optional name of technician who conducted the test}
}
\value{
A CpetMetadata S7 object
}
\description{
Stores test metadata including calibration and environmental conditions.
}
\examples{
metadata <- CpetMetadata(
  test_date = Sys.Date(),
  device = "COSMED Quark CPET",
  protocol = "Incremental ramp"
)

}
