% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{detect_cpet_format}
\alias{detect_cpet_format}
\title{Detect CPET File Format}
\usage{
detect_cpet_format(file)
}
\arguments{
\item{file}{Path to the CPET data file}
}
\value{
Character string indicating the detected format
}
\description{
Attempts to detect the metabolic cart format from file contents.
}
\keyword{internal}
