% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ExerciseQualityCriteria}
\alias{ExerciseQualityCriteria}
\title{Exercise Quality Criteria Class}
\usage{
ExerciseQualityCriteria(
  peak_rer = integer(0),
  rer_met = logical(0),
  peak_hr = integer(0),
  predicted_hr_max = integer(0),
  hr_pct_predicted = integer(0),
  hr_met = logical(0),
  vo2_plateau_detected = logical(0),
  vo2_plateau_delta = integer(0),
  plateau_met = logical(0),
  rpe_reported = integer(0),
  rpe_met = logical(0),
  lactate_mmol = integer(0),
  lactate_met = logical(0),
  criteria_met = integer(0),
  criteria_available = integer(0),
  is_maximal = logical(0),
  determination = character(0)
)
}
\arguments{
\item{peak_rer}{Peak respiratory exchange ratio achieved}

\item{rer_met}{Logical indicating if RER criterion was met (>= 1.10)}

\item{peak_hr}{Peak heart rate achieved in bpm}

\item{predicted_hr_max}{Age-predicted maximum heart rate (220 - age)}

\item{hr_pct_predicted}{Percentage of predicted max HR achieved}

\item{hr_met}{Logical indicating if HR criterion was met (>= 85\% predicted)}

\item{vo2_plateau_detected}{Logical indicating if VO2 plateau was detected}

\item{vo2_plateau_delta}{Change in VO2 (mL/min) during plateau assessment}

\item{plateau_met}{Logical indicating if plateau criterion was met (< 150 mL/min increase)}

\item{rpe_reported}{Optional reported RPE on 6-20 Borg scale}

\item{rpe_met}{Optional logical indicating if RPE criterion was met (>= 17)}

\item{lactate_mmol}{Optional blood lactate concentration in mmol/L}

\item{lactate_met}{Optional logical indicating if lactate criterion was met (>= 8.0)}

\item{criteria_met}{Number of criteria met}

\item{criteria_available}{Number of criteria that could be assessed}

\item{is_maximal}{Logical indicating if test is considered maximal (>= 3 criteria)}

\item{determination}{Character description: "maximal", "likely_maximal", "submaximal", "indeterminate"}
}
\value{
An ExerciseQualityCriteria S7 object
}
\description{
Assessment of ACSM maximal exercise criteria for determining if a CPET test
achieved true maximal effort. Based on ACSM Guidelines for Exercise Testing
and Prescription (11th edition).
}
