% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stages.R
\name{extract_stages_from_phase}
\alias{extract_stages_from_phase}
\title{Extract Stages from Phase Column}
\usage{
extract_stages_from_phase(breaths, protocol, stage_duration)
}
\arguments{
\item{breaths}{Breath data tibble}

\item{protocol}{Protocol type}

\item{stage_duration}{Stage duration in seconds}
}
\value{
Tibble with stage annotations
}
\description{
Extract Stages from Phase Column
}
\keyword{internal}
