% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stages.R
\name{extract_stages_from_power}
\alias{extract_stages_from_power}
\title{Extract Stages from Power Changes}
\usage{
extract_stages_from_power(breaths, protocol, stage_duration)
}
\arguments{
\item{breaths}{Breath data tibble}

\item{protocol}{Protocol type}

\item{stage_duration}{Stage duration for step protocols}
}
\value{
Tibble with stage annotations
}
\description{
Extract Stages from Power Changes
}
\keyword{internal}
