% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-peaks.R
\name{find_time_to_peak}
\alias{find_time_to_peak}
\title{Find Time to Peak VO2}
\usage{
find_time_to_peak(x, averaging = 30)
}
\arguments{
\item{x}{A CpetData object}

\item{averaging}{Averaging window in seconds}
}
\value{
Numeric time in seconds at peak VO2
}
\description{
Determines the time point at which peak VO2 occurred.
}
