% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{format_nutritional_state}
\alias{format_nutritional_state}
\title{Format Nutritional State}
\usage{
format_nutritional_state(state, language = "en")
}
\arguments{
\item{state}{Character: "fed" or "fasted"}

\item{language}{Language code ("en" or "fr")}
}
\value{
Formatted string
}
\description{
Format Nutritional State
}
\keyword{internal}
