% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_template_path}
\alias{get_template_path}
\title{Get Template Path}
\usage{
get_template_path(custom_template = NULL)
}
\arguments{
\item{custom_template}{Optional custom template path}
}
\value{
Path to Typst template
}
\description{
Get Template Path
}
\keyword{internal}
