% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{mod_plots_server}
\alias{mod_plots_server}
\title{Plots Module Server}
\usage{
mod_plots_server(id, language, analysis, settings = NULL)
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Reactive language value.}

\item{analysis}{Reactive CpetAnalysis object from results module.}

\item{settings}{Reactive settings list from settings module (optional).}
}
\description{
Plots Module Server
}
\keyword{internal}
