% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_results_server}
\alias{mod_results_server}
\title{Results Module Server}
\usage{
mod_results_server(
  id,
  language,
  cpet_data,
  participant,
  settings,
  prediction_source = shiny::reactive("jones"),
  dark_mode = shiny::reactive(FALSE)
)
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Reactive language value.}

\item{cpet_data}{Reactive CpetData object from upload module.}

\item{participant}{Reactive Participant object from participant module.}

\item{settings}{Reactive settings list from settings module.}

\item{prediction_source}{Reactive prediction equation source ("jones" or "prefaut").}
}
\value{
A list with reactive values:
\itemize{
\item \code{analysis}: Reactive CpetAnalysis object.
}
}
\description{
Results Module Server
}
\keyword{internal}
