% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{mod_settings_server}
\alias{mod_settings_server}
\title{Settings Module Server}
\usage{
mod_settings_server(id, language, cpet_data = shiny::reactive(NULL))
}
\arguments{
\item{id}{Module namespace ID.}

\item{language}{Reactive language value.}
}
\value{
A list with reactive values:
\itemize{
\item \code{settings}: Reactive list of analysis parameters.
}
}
\description{
Settings Module Server
}
\keyword{internal}
