% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gas_exchange}
\alias{plot_gas_exchange}
\title{Plot Gas Exchange Time Series}
\usage{
plot_gas_exchange(
  x,
  variables = c("vo2", "vco2", "rer"),
  smooth = FALSE,
  normalize = FALSE,
  language = "en"
)
}
\arguments{
\item{x}{A CpetData object}

\item{variables}{Variables to plot: any of "vo2", "vco2", "rer", "ve"}

\item{smooth}{Apply LOESS smoothing}

\item{normalize}{Normalize to body weight (show mL/kg/min for VO2)}

\item{language}{Language for labels}
}
\value{
A ggplot2 object
}
\description{
Plots VO2, VCO2, and RER against time with optional smoothing.
}
