% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power}
\alias{plot_power}
\title{Plot Power Output Response}
\usage{
plot_power(x, show_vo2 = TRUE, language = "en")
}
\arguments{
\item{x}{A CpetData object}

\item{show_vo2}{Overlay VO2 on secondary axis}

\item{language}{Language for labels}
}
\value{
A ggplot2 object
}
\description{
Plots power output against time with VO2 overlay option.
}
